+++ crates/miden-tx/src/testing/mock_chain/mod.rs
@@ -18,12 +18,10 @@ use miden_objects::{
     batch::{ProposedBatch, ProvenBatch},
     block::{
         AccountTree, AccountWitness, BlockAccountUpdate, BlockHeader, BlockInputs, BlockNoteIndex,
-        BlockNoteTree, BlockNumber, NullifierWitness, OutputNoteBatch, ProposedBlock, ProvenBlock,
+        BlockNoteTree, BlockNumber, Blockchain, NullifierTree, NullifierWitness, OutputNoteBatch,
+        ProposedBlock, ProvenBlock,
     },
-    crypto::{
-        dsa::rpo_falcon512::SecretKey,
-        merkle::{Mmr, Smt, SmtProof},
-    },
+    crypto::{dsa::rpo_falcon512::SecretKey, merkle::SmtProof},
     note::{Note, NoteHeader, NoteId, NoteInclusionProof, NoteType, Nullifier},
     testing::account_code::DEFAULT_AUTH_SCRIPT,
     transaction::{
@@ -249,13 +247,13 @@ impl PendingObjects {
 #[derive(Debug, Clone)]
 pub struct MockChain {
     /// An append-only structure used to represent the history of blocks produced for this chain.
-    chain: Mmr,
+    chain: Blockchain,
 
     /// History of produced blocks.
     blocks: Vec<ProvenBlock>,
 
     /// Tree containing the latest `Nullifier`'s tree.
-    nullifiers: Smt,
+    nullifiers: NullifierTree,
 
     /// Tree containing the latest state commitment of each account.
     account_tree: AccountTree,
@@ -282,9 +280,9 @@ pub struct MockChain {
 impl Default for MockChain {
     fn default() -> Self {
         MockChain {
-            chain: Mmr::default(),
+            chain: Blockchain::default(),
             blocks: vec![],
-            nullifiers: Smt::default(),
+            nullifiers: NullifierTree::default(),
             account_tree: AccountTree::new(),
             pending_objects: PendingObjects::new(),
             available_notes: BTreeMap::new(),
@@ -736,7 +734,7 @@ impl MockChain {
         }
 
         let block_headers = block_headers_map.values().cloned();
-        let mmr = ChainMmr::from_mmr(&self.chain, block_headers).unwrap();
+        let mmr = ChainMmr::from_blockchain(&self.chain, block_headers).unwrap();
 
         TransactionInputs::new(
             account,
@@ -877,7 +875,8 @@ impl MockChain {
             // TODO: Implement nullifier tree reset once defined at the protocol level.
             for nullifier in self.pending_objects.created_nullifiers.iter() {
                 self.nullifiers
-                    .insert(nullifier.inner(), [current_block_num.into(), ZERO, ZERO, ZERO]);
+                    .mark_spent(*nullifier, BlockNumber::from(current_block_num))
+                    .expect("nullifier should not already be spent");
             }
             let notes_tree = self.pending_objects.build_notes_tree();
 
@@ -974,7 +973,7 @@ impl MockChain {
             }
 
             self.blocks.push(block.clone());
-            self.chain.add(header.commitment());
+            self.chain.push(header.commitment());
             self.reset_pending();
 
             last_block = Some(block);
@@ -991,19 +990,19 @@ impl MockChain {
     // ACCESSORS
     // =========================================================================================
 
-    /// Returns a refernce to the current [`Mmr`] representing the blockchain.
-    pub fn block_chain(&self) -> &Mmr {
+    /// Returns a refernce to the current [`Blockchain`].
+    pub fn block_chain(&self) -> &Blockchain {
         &self.chain
     }
 
     /// Gets the latest [ChainMmr].
     pub fn latest_chain_mmr(&self) -> ChainMmr {
-        // We cannot pass the latest block as that would violate the condition in the transaction
-        // inputs that the chain length of the mmr must match the number of the reference block.
+        // We have to exclude the latest block because we need to fetch the state of the chain at
+        // that latest block, which does not include itself.
         let block_headers =
             self.blocks.iter().map(|b| b.header()).take(self.blocks.len() - 1).cloned();
 
-        ChainMmr::from_mmr(&self.chain, block_headers).unwrap()
+        ChainMmr::from_blockchain(&self.chain, block_headers).unwrap()
     }
 
     /// Creates a new [`ChainMmr`] with all reference blocks in the given iterator except for the
@@ -1017,7 +1016,7 @@ impl MockChain {
     ) -> (BlockHeader, ChainMmr) {
         let latest_block_header = self.latest_block_header().clone();
         // Deduplicate block numbers so each header will be included just once. This is required so
-        // ChainMmr::from_mmr does not panic.
+        // ChainMmr::from_blockchain does not panic.
         let reference_blocks: BTreeSet<_> = reference_blocks.into_iter().collect();
 
         // Include all block headers of the reference blocks except the latest block.
@@ -1027,7 +1026,7 @@ impl MockChain {
             .filter(|block_header| block_header.commitment() != latest_block_header.commitment())
             .collect();
 
-        let chain_mmr = ChainMmr::from_mmr(&self.chain, block_headers).unwrap();
+        let chain_mmr = ChainMmr::from_blockchain(&self.chain, block_headers).unwrap();
 
         (latest_block_header, chain_mmr)
     }
@@ -1055,8 +1054,8 @@ impl MockChain {
         let mut nullifier_proofs = BTreeMap::new();
 
         for nullifier in nullifiers {
-            let proof = self.nullifiers.open(&nullifier.inner());
-            nullifier_proofs.insert(nullifier, NullifierWitness::new(proof));
+            let witness = self.nullifiers.open(&nullifier);
+            nullifier_proofs.insert(nullifier, witness);
         }
 
         nullifier_proofs
@@ -1087,7 +1086,9 @@ impl MockChain {
 
     /// Returns a reference to the latest [`BlockHeader`].
     pub fn latest_block_header(&self) -> BlockHeader {
-        self.blocks[self.chain.forest() - 1].header().clone()
+        let chain_tip =
+            self.chain.chain_tip().expect("chain should contain at least the genesis block");
+        self.blocks[chain_tip.as_usize()].header().clone()
     }
 
     /// Gets a reference to [BlockHeader] with `block_number`.
@@ -1096,7 +1097,7 @@ impl MockChain {
     }
 
     /// Gets a reference to the nullifier tree.
-    pub fn nullifiers(&self) -> &Smt {
+    pub fn nullifiers(&self) -> &NullifierTree {
         &self.nullifiers
     }