+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - [BREAKING] Add `NetworkAccount` configuration (#1275).
 - Added support for environment variables to set up the `miden-proving-service` worker (#1281).
 - Added field identifier structs for component metadata (#1292).
+- Move `NullifierTree` and `BlockChain` from node to base (#1304).
 
 ## 0.8.2 (2025-04-18) - `miden-proving-service` crate only